requirements.txt
opencv-python
PyQt6
numpy
mss
pyautogui
Pillow
openpyxl
